pyproject.toml
[project]
name = "what-the-slack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "slack-bolt>=1.27.0",
]
